requirements.txt
# Radio PrideSync - Python Dependencies

# Core libraries
RPi.GPIO==0.7.1
smbus2==0.4.2
adafruit-circuitpython-busdevice==5.2.6

# Audio processing
pyaudio==0.2.11
pydub==0.25.1
mutagen==1.47.0

# Web interface
flask==2.3.3
flask-socketio==5.3.6

# Utilities
numpy==1.24.3
requests==2.31.0
python-dateutil==2.8.2

# Logging and configuration
colorlog==6.7.0
pyyaml==6.0.1

# Development and testing
pytest==7.4.2
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0

# Optional: For advanced audio features
scipy==1.11.2
librosa==0.10.1
